+++ index.js
@@ -3,19 +3,21 @@
  * Released under the MIT license.
  * https://opensource.org/licenses/MIT
  */
-
-'use strict';
-
 const http = require( 'http' );
 const path = require( 'path' );
 const java = require( 'java' );
-const portfinder = require( 'portfinder' );
+const portFinder = require( 'portfinder' );
 const loaderUtils = require( 'loader-utils' );
+const createLogWriter = require( './lib/log-writer' );
 
 const CONFIGURATION = 'com.realobjects.pdfreactor.Configuration';
-const PDFREACTOR = 'com.realobjects.pdfreactor.PDFreactor';
+const PDF_REACTOR = 'com.realobjects.pdfreactor.PDFreactor';
+const LEVEL = 'java.util.logging.Level';
+const LOGGER = 'java.util.logging.Logger';
+const LOGGING_HANDLER = 'com.aixigo.pdfreactor_loader.LoggingHandler';
+const LOG_WRITER = 'com.aixigo.pdfreactor_loader.LogWriter';
 
-module.exports = function (source) {
+module.exports = function( source ) {
    const options = loaderUtils.getOptions( this ) || {};
    const classpath = Array.isArray( options.classpath ) ? options.classpath : [ options.classpath ];
 
@@ -34,8 +36,9 @@ module.exports = function (source) {
    }
 
    java.ensureJvm( () => {
-      const pdfconfig = java.newInstanceSync( CONFIGURATION );
-      const pdfrenderer = java.newInstanceSync( PDFREACTOR );
+      const pdfConfig = java.newInstanceSync( CONFIGURATION );
+      const pdfRenderer = java.newInstanceSync( PDF_REACTOR );
+
       const server = http.createServer( ( req, res ) => {
          const resource = path.join( this.context, req.url.substr( 1 ) );
          this.addDependency( resource );
@@ -53,19 +56,21 @@ module.exports = function (source) {
          } );
       } );
 
-      pdfconfig.setEnableDebugModeSync( true );
-      pdfconfig.setJavaScriptModeSync( java.import( CONFIGURATION ).JavaScriptMode.ENABLED );
-      pdfconfig.setAddLinks( true );
-      pdfconfig.setAddBookmarks( true );
-      pdfconfig.setDocument( source );
+      pdfConfig.setLoggerSync( createLogger() );
+      pdfConfig.setLogLevelSync( java.import( CONFIGURATION ).LogLevel.DEBUG );
+      pdfConfig.setEnableDebugModeSync( true );
+      pdfConfig.setJavaScriptModeSync( java.import( CONFIGURATION ).JavaScriptMode.ENABLED );
+      pdfConfig.setAddLinksSync( true );
+      pdfConfig.setAddBookmarksSync( true );
+      pdfConfig.setDocumentSync( source );
 
       pipe( [
-         ( callback ) => {
+         callback => {
             listen( server, callback );
          },
          ( port, callback ) => {
-            pdfconfig.setBaseURL( 'http://127.0.0.1:' + port );
-            pdfrenderer.convertAsBinary( pdfconfig, callback );
+            pdfConfig.setBaseURL( 'http://127.0.0.1:' + port );
+            pdfRenderer.convertAsBinary( pdfConfig, callback );
          },
          ( bytes, callback ) => {
             callback( null, Buffer.from( bytes ) );
@@ -75,6 +80,28 @@ module.exports = function (source) {
          server.close();
       } );
    } );
+
+   ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+   function createLogger() {
+      const loggingHandler = java.newInstanceSync( LOGGING_HANDLER );
+      const logWriter = java.newProxy( LOG_WRITER, createLogWriter() );
+      loggingHandler.setLogWriterSync( logWriter );
+
+      const logger = java.import( LOGGER ).getLoggerSync( 'pdfreactor' );
+      // The parent handler would simply print the log messages to console without proper formatting.
+      // Hence we disable it.
+      logger.setUseParentHandlers( false );
+      logger.addHandlerSync( loggingHandler );
+      logger.setLevelSync( java.import( LEVEL ).FINEST );
+      return logger;
+   }
+};
+
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+module.exports.classpathEntries = function() {
+   return [ `${__dirname}/java` ];
 };
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -85,17 +112,17 @@ function listen( server, callback ) {
    repeat();
 
    function repeat() {
-      portfinder.getPort( (err, port) => {
-         if (err) {
+      portFinder.getPort( ( err, port ) => {
+         if( err ) {
             callback(err);
          }
          else {
             server.listen( port, err => {
-               if (err && tries-- > 0) {
+               if(err && tries-- > 0) {
                   setTimeout( repeat, 100 + (Math.random() * 100) );
                }
                else {
-                  callback(err, port);
+                  callback( err, port );
                }
             } );
          }
@@ -120,6 +147,6 @@ function pipe( fns, callback ) {
          fn.apply( null, Array.prototype.slice.call( arguments, 1, fn.length ).concat( [ iterate ] ) );
       }
    }
-};
+}