+++ java/com/aixigo/pdfreactor_loader/LogWriter.java
@@ -1,0 +1,9 @@
+package com.aixigo.pdfreactor_loader;
+
+import java.util.logging.LogRecord;
+
+public interface LogWriter {
+
+   public void writeMessage( LogRecord record );
+
+}