+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#1](https://github.com/aixigo/pdfreactor-loader/issues/1): PDFreactor log messages are now written to console 
+
 
 ## v0.2.0